much progress

+++ src/util/Vector3D.java
@@ -24,7 +24,7 @@ public class Vector3D {
         return z;
     }
 
-    public void set(int hx, int hy, int hz) {
+    public void set(double hx, double hy, double hz) {
         this.x = hx;
         this.y = hy;
         this.z = hz;
